+++ src/server/game.cpp
@@ -38,7 +38,7 @@ const size_t max_message_length = 256;
 static void truncate_message(const simple_wml::string_span& str, simple_wml::node& message) {
 	// testing for msg.size() is not sufficient but we're not getting false negatives
 	// and it's cheaper than always converting to wstring.
-	if(str.size() > chat_message::max_message_length) {
+	if(str.size() > static_cast<int>(chat_message::max_message_length)) {
 		std::string tmp(str.begin(), str.end());
 		// The string can contain utf-8 characters so truncate as wide_string otherwise
 		// a corrupted utf-8 string can be returned.
